pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ImagineIt"
version = "0.2"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "pandas",
]

[project.optional-dependencies]
gpu = [
    "diffusers",
    "transformers",
    "torch",
    "controlnet_aux",
]
beam = [
    "beam-client",
]

[project.scripts]
imagineit = "imagineit_app.main:main"

[tool.setuptools]
include-package-data = true